+++ docs/versionhistory.rst
@@ -6,6 +6,7 @@ This library adheres to `Semantic Versioning 2.0 <https://semver.org/#semantic-v
 **UNRELEASED**
 
 - Added a :pep:`302` import hook for annotating functions and classes with ``@typechecked``
+- Added a pytest plugin that activates the import hook
 - Added support for ``typing.TypedDict``
 - Deprecated ``TypeChecker`` (will be removed in v3.0)